[@osd/cross-platform] Adds cross-platform helpers (#2681)
cherry-pick from 887093d2d243045029b644680a3e8d0150318143
