+++ packages/osd-config-schema/package.json
@@ -10,6 +10,7 @@
     "osd:bootstrap": "yarn build"
   },
   "devDependencies": {
+    "@osd/cross-platform": "1.0.0",
     "typescript": "4.0.2",
     "tsd": "^0.16.0"
   },
