+++ packages/osd-cross-platform/README.md
@@ -1,0 +1,3 @@
+# `@osd/cross-platform` — OpenSearch Dashboards cross-platform helpers
+
+This package contains the helper functions to work around the differences of platforms
