+++ packages/osd-optimizer/src/optimizer/get_changes.test.ts
@@ -35,7 +35,7 @@ import path from 'path';
 jest.mock('execa');
 
 import { getChanges } from './get_changes';
-import { standardize } from '@osd/dev-utils';
+import { standardize } from '@osd/cross-platform';
 
 const execa: jest.Mock = jest.requireMock('execa');
