+++ packages/osd-plugin-helpers/src/integration_tests/build.test.ts
@@ -34,12 +34,8 @@ import Path from 'path';
 import Fs from 'fs';
 
 import execa from 'execa';
-import {
-  REPO_ROOT,
-  standardize,
-  createStripAnsiSerializer,
-  createReplaceSerializer,
-} from '@osd/dev-utils';
+import { standardize } from '@osd/cross-platform';
+import { REPO_ROOT, createStripAnsiSerializer, createReplaceSerializer } from '@osd/dev-utils';
 import extract from 'extract-zip';
 import del from 'del';
 import globby from 'globby';
