+++ packages/osd-pm/src/utils/projects_tree.ts
@@ -33,7 +33,7 @@
 import chalk from 'chalk';
 import path from 'path';
 
-import { standardize } from '@osd/utils';
+import { standardize } from '@osd/cross-platform';
 import { Project } from './project';
 
 const projectKey = Symbol('__project');
