+++ packages/osd-utils/package.json
@@ -11,6 +11,7 @@
   },
   "dependencies": {
     "@osd/config-schema": "1.0.0",
+    "@osd/cross-platform": "1.0.0",
     "load-json-file": "^6.2.0"
   },
   "devDependencies": {
